refactor(Diagram): replace parameter with `struct` field
The parameter was used to be able to write and dispatch on different
`Base.show` methods for different diagram types and their supported
output formats.

With the introduction of the intermediate `render` method this is not
necessary, nor efficient. It's simpler to include the diagram type as a
field in the `struct` and use that field directly in, amongst others,
the `render` function.