feat: make public API available without `Kroki` prefix
Apart from the string literals, the `Diagram` type and `render` function
make up the main public API of the package. The string literals are
already `export`ed, but this type and function were not fully exposed
publicly. To make it easier to support scenarios that require the use of
these explicitly, they are now `export`ed as well.