refactor(Diagram): turn inner constructor into outer constructor
This ensures the docstring attached to the constructor can be included
in the documentation. This constructor mostly exists to provide a more
straightforward API that specifies the `type` first and the
`specification` second, to accomodate for the order of the fields in the
`Diagram` type. But both orders are technically okay.