refactor(Diagram): add full keyword argument constructor
The `Diagram(::Symbol, ::AbstractString)` constructor was converted from
an inner to an outer constructor in a1de7699663d2c966afacfa0a5577784b86877b8.
The resulting `Diagram(::AbstractString, ::Symbol)` constructor was
never intended to be used and hasn't been actually shipped.

Adding an inner constructor capable of instantiating a `Diagram` from a
set of keyword arguments, is beneficial given that more (optional)
fields are to be added to the `Diagram` type.