test(string-literal): fix mermaid diagrams comparison
The intent of this test was to ensure rendering a Mermaid diagram is
equivalent when using the `mermaid` string literal or the `Diagram` type
directly. What it was testing was whether the resulting `Diagram`
instances were equivalent.

With the introduction of the `options` field in
1c814b8a9b89d3994f8337ae34723cd6980fc248, this is no longer the case as
a `Dict{String,String}()` does not equal another
`Dict{String,String}()`. This caused the test to fail. What _should be
compared is the rendered diagram.