feat: add (experimental) support for `diagramsnet`
This is considered experimental by the upstream Kroki service. At the
time of adding this support, this means it will only work with a local
instance of Kroki and the associated companion container. See
https://docs.kroki.io/kroki/setup/install/#_companion_containers.

Once the upstream Kroki service moves support for these diagrams out of
the 'experimental' phase, rendering these diagrams will just work in
general.