feat: add 'support' for rendering to JPEG
Which diagram types could be rendered to JPEG was already defined
through `LIMITED_DIAGRAM_SUPPORT`. However, no `show` method was
available so far.

The reason this adds 'support' is that the functionality to render to
JPEG appears to be broken in the Kroki service at the time of writing.
This functionality is enabled here to support the output format once the
issues upstream have been addressed, without requiring a new release.