fix(string-literal): enable escaping the string interpolation sign
The string interpolation sign (`$`) should be escapable when used within
a string literal to prevent triggering string interpolation.

This is a bit finicky, especially in the tests, as the content of a
string literal needs to be turned into a Julia string, whereas the
specification of the content will be interpreted as it would be by
Kroki. This requires careful management of escape sequences.