refactor: use a list comprehension to cover `SUPPORTED_TEXT_PLAIN_SHOW_MIME_TYPES`
It is not entirely clear why this line does not get measured as covered
by tests when the `filter` approach is used. The expectation is that
this has to do with the fact that the `filter` approach results in an
iterator being assigned to `SUPPORTED_TEXT_PLAIN_SHOW_MIME_TYPES`
resulting in a `Set`, whereas the list comprehension assigns a `Set`
constructed from a Vector`. This is pure speculation, trying to
materialize the iterator approach first does not result in the line
being covered either.