ci: detect version to replace using regular expression
In contrast to the `to` value for the
`@google/semantic-release-replace-plugin` configuration, the `from`
configuration is not interpreted as a template. Hence, it is not
possible to use the `lastRelease` variable made available by
`semantic-release`.

Instead the `to` configuration needs to be a regular expression
compatible with `new Regexp(..., "gm")`.

See
https://github.com/bauglir/Kroki.jl/runs/7515626072?check_suite_focus=true#step:3:910.