docs(service): remove dynamically generated example output from the `info` docstring
The `info` function performs HTTP requests to the configured `ENDPOINT`
to obtain an overview of the versions of the available tools for the
specific Kroki instance. Including example output for it in the
docstring of the function was a nice feature, but it hangs
precompilation on Julia v1.10 and later due to the following[^1]:

> Julia 1.10 and higher wait for all running Tasks to finish before
> writing out the precompiled (cached) version of the package. One
> consequence is that a package that launches Tasks in its __init__
> function may precompile successfully, but block precompilation of any
> packages that depend on it.

To ensure the package does not hang precompilation and can be installed
on Julia v1.10 and later the example output needs to be removed.

Fixes #48.

[^1]: https://juliatesting.github.io/Aqua.jl/dev/persistent_tasks/