test: detect theme changes based on associated fonts
For previous versions of the Kroki service, the full source of the
diagram was included in rendered results which made it possible to
directly detect the theme. This is no longer the case[^1], making it
necessary to detect the effect of a theme. The most straightforward way
to do this is by detecting fonts referenced in the rendered results.

[^1]: https://github.com/bauglir/Kroki.jl/actions/runs/8038302157/job/21954056804?pr=49#step:5:126