build(docs): migrate to `makedocs` `warnonly` keyword argument from `strict`
The `strict` keyword argument is no longer supported as of
`Documenter@1`[^1]. The `warnonly` keyword argument is its replacement.
When running in CI all errors should be critical, which is the default,
but when not they reported as warnings so they can be resolved locally
without failing the build to make it easier to make more extensive
changes to the documentation using `LiveServer` which would otherwise
error as well.

[^1]: https://github.com/JuliaDocs/Documenter.jl/blob/master/CHANGELOG.md#breaking