ci(semantic-release): register the created tag explicitly with JuliaRegistrator
In most scenarios the `main` branch is equivalent to the generated tag.
However, when signing commits the tag will be a unique object that can
be referenced instead. The tag will also be more stable/less likely to
change during the time the registration process runs, which in general
is likely not an issue but it does not hurt to guard against.