feat!: enable control over output formats used for `Base.show`
`Kroki` tries to select the most suitable output format to render a
diagram in using Julia's `Base.show` and `AbstractDisplay` capabilities.
Situations may occur where the "best" output format that gets determined
is not actually the "best" output format.

This new functionality enables control over which output formats will be
used by `Base.show`, in particular allowing specific output formats to
be disabled. This approach prevents the need for having to explicitly
construct a `Diagram` and `render`ing it and will also work when
defining diagrams using the string literals.

This is technically not a breaking change. However, given that the
overall API of this package has proven stable over a long period of time
this addition is a good point in time to move to v1.y.z.

Closes #51.